requirements.txt
Flask
requests
beautifulsoup4
python-docx
PyMuPDF
playwright
python-dotenv